x86-symbolic: Split out registers into their own module
